Rename new endpoint
