+++ backend/canisters/community/api/src/main.rs
@@ -47,7 +47,7 @@ fn main() {
     generate_candid_method!(community, remove_reaction, update);
     generate_candid_method!(community, reset_invite_code, update);
     generate_candid_method!(community, send_message, update);
-    generate_candid_method!(community, set_user_display_name, update);
+    generate_candid_method!(community, set_member_display_name, update);
     generate_candid_method!(community, toggle_mute_notifications, update);
     generate_candid_method!(community, unblock_user, update);
     generate_candid_method!(community, undelete_messages, update);
