+++ backend/canisters/community/api/src/updates/mod.rs
@@ -35,7 +35,7 @@ pub mod remove_member_from_channel;
 pub mod remove_reaction;
 pub mod reset_invite_code;
 pub mod send_message;
-pub mod set_user_display_name;
+pub mod set_member_display_name;
 pub mod toggle_mute_notifications;
 pub mod unblock_user;
 pub mod undelete_messages;
